more work on cleanup - still ugly
